requirements.txt
PySide6>=6.5.0
opencv-python>=4.8.0
mediapipe>=0.10.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
bcrypt>=4.0.0
cryptography>=41.0.0
pyperclip>=1.8.2
PyInstaller>=6.0.0
